requirements.txt
openai==1.12.0
python-dotenv==1.0.0
streamlit==1.32.0
# googlemaps==4.10.0
requests==2.31.0
python-dateutil==2.8.2 